dev: extract function for parsing package spec
